fix: use float output dtype in Tabulated1D.__call__ to avoid integer truncation
np.zeros_like(x) inherits the input array's dtype.  When an integer-valued
array is passed, interpolated float values are silently truncated to int
(e.g. f(5)=0.5 but f(np.array([5]))=[0]).  Initialize with dtype=float,
matching the same fix already applied to sum_functions in the prior commit.

Adds test_tabulated1d_integer_input covering the exact truncation scenario
reported by @dylanpulver.

+++ openmc/data/function.py
@@ -177,8 +177,10 @@ class Tabulated1D(Function1D):
 
         x = np.array(x)
 
-        # Create output array
-        y = np.zeros_like(x)
+        # Create output array.  Use a floating-point dtype so that
+        # interpolated values are not truncated when the input is an
+        # integer-valued array.
+        y = np.zeros_like(x, dtype=float)
 
         # Get indices for interpolation
         idx = np.searchsorted(self.x, x, side='right') - 1
